requirements.txt
django
efficientnet-pytorch
torch 
torchvision 
pillow
gunicorn
whitenoise
python-dotenv
django-environ
tqdm
opencv-python